合并最新代码

+++ api/source/mk_thread.cpp
@@ -1,7 +1,7 @@
 ﻿/*
  * Copyright (c) 2016 The ZLMediaKit project authors. All Rights Reserved.
  *
- * This file is part of ZLMediaKit(https://github.com/xiongziliang/ZLMediaKit).
+ * This file is part of ZLMediaKit(https://github.com/xia-chu/ZLMediaKit).
  *
  * Use of this source code is governed by MIT license that can be found in the
  * LICENSE file in the root of the source tree. All contributing project authors
@@ -77,7 +77,7 @@ public:
         _task->cancel();
     }
 
-    void start(int ms ,EventPoller &poller){
+    void start(uint64_t ms ,EventPoller &poller){
         weak_ptr<TimerForC> weak_self = shared_from_this();
         _task = poller.doDelayTask(ms, [weak_self]() {
             auto strong_self = weak_self.lock();
